worker/requirements.txt
amqp==5.3.1
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
flake8==7.2.0
idna==3.10
kombu==5.5.4
mccabe==0.7.0
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
prompt_toolkit==3.0.51
pycodestyle==2.13.0
pyflakes==3.3.2
pymongo==4.13.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
